images/psql-init/requirements.txt
psycopg2-binary==2.9.3
sqlalchemy==1.3.5  
